requirements.txt
PyYAML==5.3.1
colorama==0.4.3